+++ src/math_stuff.cu
@@ -1,9 +1,9 @@
 #include "math_stuff.h"
 
-__int64 euclidean_gcd(__int64 a_in, __int64 b_in) {
-	__int64 a = a_in;
-	__int64 b = b_in;
-	__int64 t, m;
+long long int euclidean_gcd(long long int a_in, long long int b_in) {
+	long long int a = a_in;
+	long long int b = b_in;
+	long long int t, m;
 
 	if (b > a) {
 		t = a; a = b; b = t;
@@ -18,9 +18,9 @@ __int64 euclidean_gcd(__int64 a_in, __int64 b_in) {
 	return a;
 }
 
-__int64 power_mod(__int64 base, __int64 exponent, __int64 modulus) {
-	__int64 i;
-	__int64 number = base;
+long long int power_mod(long long int base, long long int exponent, long long int modulus) {
+	long long int i;
+	long long int number = base;
 
 	for(i = 1; i < exponent; i++) {
 		number = (number * base) % modulus;
